requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
django-ckeditor==6.5.1
django-js-asset==2.0.0
pycodestyle==2.10.0
python-decouple==3.8
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
whitenoise==6.4.0
